requirements.txt
Django==4.2.9
django-model-utils==4.2.0
djangorestframework==3.14.0
django-filter==23.5
